Move casualty details to the home widget settings
